Cleaned up code massively now that direction is known
