+++ Gruntfile.js
@@ -15,9 +15,8 @@ module.exports = function (grunt) {
         jshint: {
             all: [
                 'Gruntfile.js',
-                'src/*.js',
-                'test/*.js',
-                '<%= nodeunit.tests %>'
+                'lib/**/*.js',
+                'test/**/*.js'
             ],
             options: {
                 jshintrc: '.jshintrc'
@@ -26,22 +25,7 @@ module.exports = function (grunt) {
 
         clean: {
             dist: {
-                src: ['dist/']
-            }
-        },
-
-        // copy all the source files we're going to combine to a temporary directory
-        copy: {
-            src: {
-                cwd: 'src',
-                src: '**',
-                dest: 'dist'
-            },
-            utils: {
-                cwd: 'bower_components/kaazing-client-javascript-util',
-                src: '**',
-                dest: 'dist/tmp/utils',
-                expand: true
+                src: ['build/']
             }
         },
 
@@ -64,17 +48,12 @@ module.exports = function (grunt) {
             }
         },
 
-        // Tests for unit tests
-        nodeunit: {
-            tests: ['test/*test.js', 'test/*Test.js']
-        },
-
         mochaTest: {
             testBase: {
                 options: {
                     reporter: 'spec',
-                    require: 'src/testFrameworks/mocha-k3po.js',
-                    //captureFile : "tobedecided.txt"
+                    require: 'lib/testFrameworks/mocha-k3po.js',
+                    captureFile : "build/mochaTest.txt",
                     quiet:false
                 },
                 src: ['test/base/*spec.js']
@@ -83,36 +62,17 @@ module.exports = function (grunt) {
                 options: {
                     reporter: 'spec',
                     ui: 'mocha-k3po',
-                    require: 'src/testFrameworks/mocha-k3po.js',
-                    //captureFile : "tobedecided.txt"
-                    quiet:false,
-                    hmm:"what"
+                    require: 'lib/testFrameworks/mocha-k3po.js',
+                    captureFile : "build/testMochaK3po.txt",
+                    quiet:false
                 },
                 src: ['test/testFrameworks/mocha-k3po*.js']
             }
         },
 
-        connect: {
-            server: {
-                options: {
-                    port: 8080,
-                    hostname: 'localhost',
-                    base: "test/web",
-                    keepalive: false
-                }
-            }
-        },
-
-        browserify: {
-            vendor: {
-                src: ["src/base/**/*.js", "scr/testFrameworks/**/*.js", "src/web/BboshTransportFactory.js"],
-                dest: 'public/k3po.js'
-                //options: {
-                //    require: ['jquery'],
-                //    alias: {
-                //        momentWrapper: './lib/moments.js'
-                //    }
-                //}
+        webdriver: {
+            test: {
+                configFile: './wdio.conf.js'
             }
         }
     });
@@ -122,19 +82,11 @@ module.exports = function (grunt) {
     grunt.loadNpmTasks('grunt-contrib-clean');
     grunt.loadNpmTasks('grunt-webdriver');
     grunt.loadNpmTasks('grunt-k3po');
-    grunt.loadNpmTasks('grunt-contrib-nodeunit');
     grunt.loadNpmTasks('grunt-mocha-test');
-    grunt.loadNpmTasks('grunt-contrib-connect');
-    grunt.loadNpmTasks('grunt-contrib-connect');
-    grunt.loadNpmTasks('grunt-browserify');
-    // DPW - Current working setup is running the following 3 tasks in two teminals
     grunt.registerTask('runRobot', ['k3po:daemon']);
     grunt.registerTask('startServer', ['connect']);
-    grunt.registerTask('testBBosh', ['mochaTest']);
-    grunt.registerTask('browser', ['browserify']);
-    grunt.loadNpmTasks('grunt-contrib-concat');
+    grunt.loadNpmTasks('grunt-webdriver');
 
     // DPW - This does not work currently becuase of bugs stopping the robot, and stopping nodeunit
-    // grunt.registerTask('default', ['jshint', 'clean', 'k3po:start', 'connect', 'nodeunit', 'mochaTest', 'k3po:stop']);
-    grunt.registerTask('default', ['jshint', 'clean', 'mochaTest:testBase', 'k3po:start', 'mochaTest:testMochaK3po', 'browser']);
+    grunt.registerTask('default', ['clean', 'jshint',  'mochaTest:testBase', 'k3po:start', 'mochaTest:testMochaK3po']);
 };
