+++ lib/K3poControl.js
@@ -1,204 +1,23 @@
 'use strict';
+var ErrorEvent = require('./events/ErrorEvent.js');
+var FinishedEvent = require('./events/FinishedEvent.js');
+var NotifiedEvent = require('./events/NotifiedEvent.js');
+var PreparedEvent = require('./events/PreparedEvent.js');
+var StartedEvent = require('./events/StartedEvent.js');
+var Command = require('./commands/Command.js');
 
-var controlTransportFactory = require("../../src/base/ControlTransportFactory.js");
-
-/**
- * Command is an abstract class that represents a K3po Command
- * @param type
- * @constructor
- */
-function Command(type) {
-    this.type = type;
-}
-
-Command.prototype.getType = function () {
-    return this.type;
-};
+var controlTransportFactory = require("./transports/ControlTransportFactory.js");
 
 /**
- * Aborts the script execution
+ * K3poControl is a class that connects to K3poControl
+ * @param connectURL
  * @constructor
  */
-function AbortCommand() {
-    Command.call(this, "ABORT");
+function K3poControl() {
+    this.connection = null;
+    this.queuedEvents = [];
+    this.eventCallbacks = [];
 }
-
-AbortCommand.prototype = Object.create(Command.prototype);
-
-AbortCommand.prototype.constructor = AbortCommand;
-
-/**
- * Awaits the barrier
- * @param barrier
- * @constructor
- */
-function AwaitCommand(barrier) {
-    Command.call(this, "AWAIT");
-    this.barrier = barrier;
-}
-
-AwaitCommand.prototype = Object.create(Command.prototype);
-
-AwaitCommand.prototype.constructor = AwaitCommand;
-
-AwaitCommand.prototype.getBarrier = function () {
-    return this.barrier;
-};
-
-/**
- * Notify k3po that a barrier has been triggered
- * @param barrier
- * @constructor
- */
-function NotifyCommand(barrier) {
-    Command.call(this, "NOTIFY");
-    this.barrier = barrier;
-}
-
-NotifyCommand.prototype = Object.create(Command.prototype);
-
-NotifyCommand.prototype.constructor = NotifyCommand;
-
-NotifyCommand.prototype.getBarrier = function () {
-    return this.barrier;
-};
-
-/**
- * Prepares the driver for running
- * @param scripts
- * @constructor
- */
-function PrepareCommand(scripts) {
-    Command.call(this, "PREPARE");
-
-    if (scripts != null && scripts.constructor === Array) {
-        this.scripts = scripts;
-    } else {
-        throw "Invalid argument, scripts is not an Array";
-    }
-}
-
-PrepareCommand.prototype = Object.create(Command.prototype);
-
-PrepareCommand.prototype.constructor = PrepareCommand;
-
-PrepareCommand.prototype.getScripts = function () {
-    return this.scripts;
-};
-
-/**
- * Starts the script execution
- * @constructor
- */
-function StartCommand() {
-    Command.call(this, "START");
-}
-
-StartCommand.prototype = Object.create(Command.prototype);
-
-StartCommand.prototype.constructor = StartCommand;
-
-/**
- * Event is an abstract class that represents a K3po Event
- * @constructor
- */
-function Event(type) {
-    this.type = type;
-}
-
-Event.prototype.getType = function () {
-    return this.type;
-};
-
-/**
- * When an error has occured
- * @param summary
- * @param description
- * @constructor
- */
-function ErrorEvent(summary, description) {
-    Event.call(this, "ERROR");
-    this.summary = summary;
-    this.description = description;
-}
-
-ErrorEvent.prototype = Object.create(Event.prototype);
-
-ErrorEvent.prototype.constructor = ErrorEvent;
-
-ErrorEvent.prototype.getSummary = function () {
-    return this.summary;
-};
-
-ErrorEvent.prototype.getDescription = function () {
-    return this.description;
-};
-
-/**
- * When script execution has finished
- * @param script
- * @constructor
- */
-function FinishedEvent(script) {
-    Event.call(this, "FINISHED");
-    this.script = script;
-}
-
-FinishedEvent.prototype = Object.create(Event.prototype);
-
-FinishedEvent.prototype.constructor = FinishedEvent;
-
-FinishedEvent.prototype.getScript = function () {
-    return this.script;
-};
-
-/**
- * Notification that a barrier has been notified
- * @param barrier
- * @constructor
- */
-function NotifiedEvent(barrier) {
-    Event.call(this, "NOTIFIED");
-    this.barrier = barrier;
-}
-
-NotifiedEvent.prototype = Object.create(Event.prototype);
-
-NotifiedEvent.prototype.constructor = NotifiedEvent;
-
-NotifiedEvent.prototype.getBarrier = function () {
-    return this.barrier;
-};
-
-/**
- * Notification that the script is prepared
- * @param script
- * @constructor
- */
-function PreparedEvent(script) {
-    Event.call(this, "PREPARED");
-    this.script = script;
-}
-
-PreparedEvent.prototype = Object.create(Event.prototype);
-
-PreparedEvent.prototype.constructor = PreparedEvent;
-
-PreparedEvent.prototype.getScript = function () {
-    return this.script;
-};
-
-/**
- * Notification that the script has started
- * @constructor
- */
-function StartedEvent() {
-    Event.call(this, "STARTED");
-}
-
-StartedEvent.prototype = Object.create(Event.prototype);
-
-StartedEvent.prototype.constructor = StartedEvent;
 
 /**
  * Sends a Command to the K3poDriver
@@ -253,7 +72,6 @@ K3poControl.prototype.on = function (event, listener) {
         case "NOTIFIED":
         case "PREPARED":
         case "STARTED":
-            // TODO, as of know there is only one listener per event type
             this.eventCallbacks[event] = listener;
             break;
         default:
@@ -269,17 +87,6 @@ K3poControl.prototype.on = function (event, listener) {
 };
 
 K3poControl.prototype.addEventListener = K3poControl.prototype.on;
-
-/**
- * K3poControl is a class that connects to K3poControl
- * @param connectURL
- * @constructor
- */
-function K3poControl() {
-    this.connection = null;
-    this.queuedEvents = [];
-    this.eventCallbacks = [];
-}
 
 /**
  * Connects to the K3po Server
@@ -341,7 +148,7 @@ K3poControl.prototype.connect = function (connectURL, callback) {
     }
 
     function parseStarted() {
-        return new StartedEvent()
+        return new StartedEvent();
     }
 
     var _this = this;
@@ -391,11 +198,4 @@ K3poControl.prototype.disconnect = function (callback) {
     this.connection.disconnect(callback);
 };
 
-var k3poControl = exports;
-k3poControl.K3poControl = K3poControl;
-k3poControl.AbortCommand = AbortCommand;
-k3poControl.AwaitCommand = AwaitCommand;
-k3poControl.NotifyCommand = NotifyCommand;
-k3poControl.PrepareCommand = PrepareCommand;
-k3poControl.StartCommand = StartCommand;
-// Don't think there is a need to exports events because only K3poControl Constructs them
+module.exports = K3poControl;
