+++ lib/K3poRunner.js
@@ -1,10 +1,9 @@
-var k3poControl = require('../../src/base/K3poControl.js'),
-    K3poControl = k3poControl.K3poControl,
-    AbortCommand = k3poControl.AbortCommand,
-    AwaitCommand = k3poControl.AwaitCommand,
-    NotifyCommand = k3poControl.NotifyCommand,
-    PrepareCommand = k3poControl.PrepareCommand,
-    StartCommand = k3poControl.StartCommand;
+var K3poControl = require('./K3poControl.js');
+var AbortCommand = require('./commands/AbortCommand.js');
+var AwaitCommand = require('./commands/AwaitCommand.js');
+var NotifyCommand = require('./commands/NotifyCommand.js');
+var PrepareCommand = require('./commands/PrepareCommand.js');
+var StartCommand = require('./commands/StartCommand.js');
 
 function K3poRunner(k3poControl, done) {
     this.state = "INITIAL"; // INITIAL, START, STARTED
@@ -98,9 +97,9 @@ K3poRunner.prototype.start = function (callback) {
     }
 };
 
-K3poRunner.prototype.abort = function(callback){
-    if(!callback){
-        callback = function(){
+K3poRunner.prototype.abort = function (callback) {
+    if (!callback) {
+        callback = function () {
 
         };
     }
