+++ lib/events/Event.js
@@ -1,0 +1,13 @@
+/**
+ * Event is an abstract class that represents a K3po Event
+ * @constructor
+ */
+function Event(type) {
+    this.type = type;
+}
+
+Event.prototype.getType = function () {
+    return this.type;
+};
+
+module.exports = Event;
