+++ lib/testFrameworks/mocha-k3po.js
@@ -1,10 +1,10 @@
 var Mocha = require('mocha');
-Suite = require('mocha/lib/suite'),
+var Suite = require('mocha/lib/suite'),
     Test = require('mocha/lib/test'),
     escapeRe = require('escape-string-regexp');
 
-var K3poRunner = require('../../src/testFrameworks/K3poRunner.js');
-var K3poControl = require('../../src/base/K3poControl.js').K3poControl;
+var K3poRunner = require('../K3poRunner.js');
+var K3poControl = require('../K3poControl.js');
 
 var assert = require('assert');
 
@@ -139,6 +139,6 @@ module.exports = Mocha.interfaces['mocha-k3po'] = function (suite) {
 
         context.awaitBarrier = function (barrierName, callback) {
             k3poRunner.awaitBarrier(barrierName, callback);
-        }
+        };
     });
 };
