+++ lib/transports/ControlTransportApi.js
@@ -48,4 +48,4 @@ ControlTransportApi.prototype.onMessage = function (callback) {
     throw "onMessage(callback) Not Implemented";
 };
 
-exports.ControlTransportApi = ControlTransportApi;
+module.exports = ControlTransportApi;
