+++ lib/transports/ControlTransportFactory.js
