+++ lib/transports/ControlTransportFactorySpi.js
@@ -7,7 +7,7 @@ ControlTransportFactorySpi.prototype.getScheme = function () {
 };
 
 ControlTransportFactorySpi.prototype.connect = function (url, callback) {
-    throw "connect(url, callback) not implemented"
+    throw "connect(url, callback) not implemented";
 };
 
-exports.ControlTransportFactorySpi = ControlTransportFactorySpi;
+module.exports = ControlTransportFactorySpi;
