+++ lib/transports/TcpTransportFactory.js
@@ -1,7 +1,7 @@
 var net = require('net');
-var ControlTransportApi = require('../base/ControlTransportApi.js').ControlTransportApi;
-var ControlTransportFactorySpi = require('../base/ControlTransportFactorySpi.js').ControlTransportFactorySpi;
-var controlTransportFactory = require('../base/ControlTransportFactory.js');
+var ControlTransportApi = require('./ControlTransportApi.js');
+var ControlTransportFactorySpi = require('./ControlTransportFactorySpi.js');
+var controlTransportFactory = require('./ControlTransportFactory.js');
 
 /**
  * K3po Control Transport for when running in node
