--- src/web/BboshTransportFactory.js
@@ -1,9 +1,0 @@
-/**
- * K3po Control Transport for when running in the browser
- * @constructor
- */
-function BboshTransport(){
-    // TODO
-}
-
-module.export = BboshTransportFactory;
