--- test-old/web/bbosh-test.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <script src="Bbosh.js"></script>
-</head>
-<body id="result">
-
-</body>
-</html>
